requirements.txt
fpdf
Jinja2>=3.1.6
Keras
h5py
imageio
librosa
matplotlib
mock
mypy
numpy==2.2.6
olefile
pandas
Pillow
PyYAML
pydub
pylance
scikit-learn
scipy
seaborn
timm
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
torchmetrics
tqdm